Update README.md
This program is a simple example of a recursive binary search and is lesson 3 in a study of algorithmic thinking.
This program takes a sorted range, determines the midpoint, ignores the portion below or above the midpoint, thus
creating a new  list.  The program continues recursively calling the function "recusive_biary_search " until it
reaches its target, displaying true (number in the list found)or false (number in the list not found).